Move collection content to the separate component
